[refactor] #100 enum 위치 및 요소 수정
