+++ src/main/java/nova/backend/domain/challenge/entity/Challenge.java
@@ -42,22 +42,30 @@ public class Challenge extends BaseTimeEntity {
     @Column(nullable = false)
     private int successCount = 10;
 
+    /**
+     * 챌린지 자연스러운 중단 상태 자체
+     */
+    public enum ChallengeStatus {
+        IN_PROGRESS, // 아직 완료되지 않음
+        COMPLETED,   // 성공 기준 도달
+        CANCELED     // 중단됨 (참여를 중단한 상태)
+    }
+
     public int getParticipantCount() {
         return participations.size();
     }
 
     public int getCompletedCount() {
         return (int) participations.stream()
-                .filter(p -> p.getChallengeStatus() == ParticipationStatus.COMPLETED)
+                .filter(p -> p.getChallengeStatus() == ChallengeStatus.COMPLETED)
                 .count();
     }
 
     public int getCanceledCount() {
         return (int) participations.stream()
-                .filter(p -> p.getChallengeStatus() == ParticipationStatus.CANCELED)
+                .filter(p -> p.getChallengeStatus() == ChallengeStatus.CANCELED)
                 .count();
     }
-
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "cafe_id")
@@ -65,5 +73,6 @@ public class Challenge extends BaseTimeEntity {
 
     @OneToMany(mappedBy = "challenge", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<ChallengeParticipation> participations = new ArrayList<>();
+
 }
