--- src/main/java/nova/backend/domain/challenge/entity/ChallengeStatus.java
@@ -1,7 +1,0 @@
-package nova.backend.domain.challenge.entity;
-
-public enum ChallengeStatus {
-    IN_PROGRESS, // 아직 완료되지 않음
-    COMPLETED    // 성공 기준 도달
-}
-
